#007: Created a new subclass of JSONClient, AuthenticatedJSONClient, and moved all of the authentication-related functionality there.
